Add clarification on IAM credentials for ssosync
"Ensure that the default credentials are for the AWS account you delegated admin for AWS IAM Identity Center to. This should be credntials for permissions limited IAM User, it can't be a user in AWS IAM Identity Center since the sync process is going to make changes. The cloudformation template (template.yaml) in this repository includes an IAM policy that grants the minimal set of rights for SSOSync to function.."