feat: sync-method groups filter by include groups as well
In this commit we extend the SyncUsersGroups method
to filter groups using the --include-groups flag.

A group is included only if the e-mail elements of
the include-groups flag match the e-mail elements of
google groups.

The filtering is done after the ignoreGroup method
is executed. This way we ensure that ignore takes
priority.