200 name handling (#204)
Google directory will accept a 'zero width space' for an empty name but will not accept a 'space', conversely AWS IAM Identity Store will accept a 'space' for an empty name but not a 'zero width space'. This change replaces 'zero width space' characters with a 'space' in the name fields of users fetched from the Google directory.