Restrict Maximum Concurrency for the ssosync lambda to 1 #207
Issue #, if available:
#206

Description of changes:
Updated the CloudFormation template to hardcode the maximum allowed concurrency for the lambda function to 1. There is no situation under which it would be safe to run more than one instance.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.