Add a --dry-run feature
Fixes #107

Instead of branching on every action in sync.go (which is cumbersome,
not future-proof and potentially error-prone) this implementation
creates a fake with pass-through methods for AWS clients.

For lookups the actions are passed-through* to the original client,
however the actions are implemented as no-ops.

This solutions is:
- reasonably correct -> there isn't a possibility of missing an action
  call [ to the faked AWS clients ] which would violate the premise of
  dry run
- future-proof -> the above applies to the calls that will be added in
the future

*small caveat with user creation - there's a virtual map to which the
users are added on creation and can be looked up from on subsequent
actions. Without this contraption a dry-run is just not practical
because it often fails due to ErrUserNotFound